Update provided by Tony Canonico for June 2017
We have also reached out to Columbia University, a leading research and treatment facility 
for Lyme disease, hoping they can aid in the treatment of his symptoms and co-infections. Lyme Disease is becoming a widespread epidemic in the Tri-state area, and it is a disease that has many 
potential complications for secondary infections if not treated quickly. With your help, 
hopefully Nicholas can beat this horribly misdiagnosed, debilitating disease.